refactor(llmproxy): complete approval body editor consolidation
